+++ src/parser/inline/media.test.ts
@@ -83,8 +83,10 @@ describe('Unit: parser/inline/media', () => {
 
     it('attribute', () => {
       assert.deepStrictEqual(inspect(parser('![]{/ nofollow}')), [['<a href="/" rel="nofollow noreferrer" target="_blank"><img class="media" data-src="/" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('![]{/ aspect-ratio}')), [['<a href="/" target="_blank"><img class="media invalid" data-src="/" alt=""></a>'], '']);
       assert.deepStrictEqual(inspect(parser('![]{/ constructor}')), [['<a href="/" class="invalid" target="_blank"><img class="media invalid" data-src="/" alt=""></a>'], '']);
       assert.deepStrictEqual(inspect(parser('!{/ nofollow}')), [['<a href="/" rel="nofollow noreferrer" target="_blank"><img class="media" data-src="/" alt=""></a>'], '']);
+      assert.deepStrictEqual(inspect(parser('!{/ aspect-ratio="4/3"}')), [['<a href="/" target="_blank"><img class="media" data-src="/" alt="" aspect-ratio="4/3"></a>'], '']);
     });
 
   });